Split CLI into focused submodules
- Extract auth, config, project, and server CLI commands
- Keep `cli.ts` as the top-level command wiring entrypoint
