+++ apps/server/src/cli/auth.ts
@@ -1,0 +1,243 @@
+import { AuthSessionId } from "@t3tools/contracts";
+import { Console, Effect, Layer, Option, References } from "effect";
+import { Argument, Command, Flag, GlobalFlag } from "effect/unstable/cli";
+
+import { AuthControlPlaneRuntimeLive } from "../auth/Layers/AuthControlPlane.ts";
+import { AuthControlPlane } from "../auth/Services/AuthControlPlane.ts";
+import type { AuthControlPlaneShape } from "../auth/Services/AuthControlPlane.ts";
+import {
+  formatIssuedPairingCredential,
+  formatIssuedSession,
+  formatPairingCredentialList,
+  formatSessionList,
+} from "../cliAuthFormat.ts";
+import { ServerConfig } from "../config.ts";
+import {
+  authLocationFlags,
+  type CliAuthLocationFlags,
+  DurationFromString,
+  resolveCliAuthConfig,
+} from "./config.ts";
+
+const runWithAuthControlPlane = <A, E>(
+  flags: CliAuthLocationFlags,
+  run: (authControlPlane: AuthControlPlaneShape) => Effect.Effect<A, E>,
+  options?: {
+    readonly quietLogs?: boolean;
+  },
+) =>
+  Effect.gen(function* () {
+    const logLevel = yield* GlobalFlag.LogLevel;
+    const config = yield* resolveCliAuthConfig(flags, logLevel);
+    const minimumLogLevel = options?.quietLogs ? "Error" : config.logLevel;
+    return yield* Effect.gen(function* () {
+      const authControlPlane = yield* AuthControlPlane;
+      return yield* run(authControlPlane);
+    }).pipe(
+      Effect.provide(
+        Layer.mergeAll(AuthControlPlaneRuntimeLive).pipe(
+          Layer.provide(Layer.succeed(ServerConfig, config)),
+          Layer.provide(Layer.succeed(References.MinimumLogLevel, minimumLogLevel)),
+        ),
+      ),
+    );
+  });
+
+const ttlFlag = Flag.string("ttl").pipe(
+  Flag.withSchema(DurationFromString),
+  Flag.withDescription("TTL, for example `5m`, `1h`, `30d`, or `15 minutes`."),
+  Flag.optional,
+);
+
+const jsonFlag = Flag.boolean("json").pipe(
+  Flag.withDescription("Emit JSON instead of human-readable output."),
+  Flag.withDefault(false),
+);
+
+const sessionRoleFlag = Flag.choice("role", ["owner", "client"]).pipe(
+  Flag.withDescription("Role for the issued bearer session."),
+  Flag.withDefault("owner"),
+);
+
+const labelFlag = Flag.string("label").pipe(
+  Flag.withDescription("Optional human-readable label."),
+  Flag.optional,
+);
+
+const subjectFlag = Flag.string("subject").pipe(
+  Flag.withDescription("Optional session subject."),
+  Flag.optional,
+);
+
+const baseUrlFlag = Flag.string("base-url").pipe(
+  Flag.withDescription("Optional public base URL used to print a ready `/pair#token=...` link."),
+  Flag.optional,
+);
+
+const tokenOnlyFlag = Flag.boolean("token-only").pipe(
+  Flag.withDescription("Print only the issued bearer token."),
+  Flag.withDefault(false),
+);
+
+const pairingCreateCommand = Command.make("create", {
+  ...authLocationFlags,
+  ttl: ttlFlag,
+  label: labelFlag,
+  baseUrl: baseUrlFlag,
+  json: jsonFlag,
+}).pipe(
+  Command.withDescription("Issue a new client pairing token."),
+  Command.withHandler((flags) =>
+    runWithAuthControlPlane(
+      flags,
+      (authControlPlane) =>
+        Effect.gen(function* () {
+          const issued = yield* authControlPlane.createPairingLink({
+            role: "client",
+            subject: "one-time-token",
+            ...(Option.isSome(flags.ttl) ? { ttl: flags.ttl.value } : {}),
+            ...(Option.isSome(flags.label) ? { label: flags.label.value } : {}),
+          });
+          const output = formatIssuedPairingCredential(issued, {
+            json: flags.json,
+            ...(Option.isSome(flags.baseUrl) ? { baseUrl: flags.baseUrl.value } : {}),
+          });
+          yield* Console.log(output);
+        }),
+      {
+        quietLogs: flags.json,
+      },
+    ),
+  ),
+);
+
+const pairingListCommand = Command.make("list", {
+  ...authLocationFlags,
+  json: jsonFlag,
+}).pipe(
+  Command.withDescription("List active client pairing tokens without revealing their secrets."),
+  Command.withHandler((flags) =>
+    runWithAuthControlPlane(
+      flags,
+      (authControlPlane) =>
+        Effect.gen(function* () {
+          const pairingLinks = yield* authControlPlane.listPairingLinks({ role: "client" });
+          yield* Console.log(formatPairingCredentialList(pairingLinks, { json: flags.json }));
+        }),
+      {
+        quietLogs: flags.json,
+      },
+    ),
+  ),
+);
+
+const pairingRevokeCommand = Command.make("revoke", {
+  ...authLocationFlags,
+  id: Argument.string("id").pipe(Argument.withDescription("Pairing credential id to revoke.")),
+}).pipe(
+  Command.withDescription("Revoke an active client pairing token."),
+  Command.withHandler((flags) =>
+    runWithAuthControlPlane(flags, (authControlPlane) =>
+      Effect.gen(function* () {
+        const revoked = yield* authControlPlane.revokePairingLink(flags.id);
+        yield* Console.log(
+          revoked
+            ? `Revoked pairing credential ${flags.id}.\n`
+            : `No active pairing credential found for ${flags.id}.\n`,
+        );
+      }),
+    ),
+  ),
+);
+
+const pairingCommand = Command.make("pairing").pipe(
+  Command.withDescription("Manage one-time client pairing tokens."),
+  Command.withSubcommands([pairingCreateCommand, pairingListCommand, pairingRevokeCommand]),
+);
+
+const sessionIssueCommand = Command.make("issue", {
+  ...authLocationFlags,
+  ttl: ttlFlag,
+  role: sessionRoleFlag,
+  label: labelFlag,
+  subject: subjectFlag,
+  tokenOnly: tokenOnlyFlag,
+  json: jsonFlag,
+}).pipe(
+  Command.withDescription("Issue a bearer session token for headless or remote clients."),
+  Command.withHandler((flags) =>
+    runWithAuthControlPlane(
+      flags,
+      (authControlPlane) =>
+        Effect.gen(function* () {
+          const issued = yield* authControlPlane.issueSession({
+            role: flags.role,
+            ...(Option.isSome(flags.ttl) ? { ttl: flags.ttl.value } : {}),
+            ...(Option.isSome(flags.label) ? { label: flags.label.value } : {}),
+            ...(Option.isSome(flags.subject) ? { subject: flags.subject.value } : {}),
+          });
+          yield* Console.log(
+            formatIssuedSession(issued, {
+              json: flags.json,
+              tokenOnly: flags.tokenOnly,
+            }),
+          );
+        }),
+      {
+        quietLogs: flags.json || flags.tokenOnly,
+      },
+    ),
+  ),
+);
+
+const sessionListCommand = Command.make("list", {
+  ...authLocationFlags,
+  json: jsonFlag,
+}).pipe(
+  Command.withDescription("List active sessions without revealing bearer tokens."),
+  Command.withHandler((flags) =>
+    runWithAuthControlPlane(
+      flags,
+      (authControlPlane) =>
+        Effect.gen(function* () {
+          const sessions = yield* authControlPlane.listSessions();
+          yield* Console.log(formatSessionList(sessions, { json: flags.json }));
+        }),
+      {
+        quietLogs: flags.json,
+      },
+    ),
+  ),
+);
+
+const sessionRevokeCommand = Command.make("revoke", {
+  ...authLocationFlags,
+  sessionId: Argument.string("session-id").pipe(
+    Argument.withDescription("Session id to revoke."),
+    Argument.withSchema(AuthSessionId),
+  ),
+}).pipe(
+  Command.withDescription("Revoke an active session."),
+  Command.withHandler((flags) =>
+    runWithAuthControlPlane(flags, (authControlPlane) =>
+      Effect.gen(function* () {
+        const revoked = yield* authControlPlane.revokeSession(flags.sessionId);
+        yield* Console.log(
+          revoked
+            ? `Revoked session ${flags.sessionId}.\n`
+            : `No active session found for ${flags.sessionId}.\n`,
+        );
+      }),
+    ),
+  ),
+);
+
+const sessionCommand = Command.make("session").pipe(
+  Command.withDescription("Manage bearer sessions."),
+  Command.withSubcommands([sessionIssueCommand, sessionListCommand, sessionRevokeCommand]),
+);
+
+export const authCommand = Command.make("auth").pipe(
+  Command.withDescription("Manage the local auth control plane for headless deployments."),
+  Command.withSubcommands([pairingCommand, sessionCommand]),
+);
